requirements/dev.txt
-r main.txt
coverage>=4.2
django-factory>=0.11
django-debug-toolbar
factory-boy>=2.4
fake-factory>=0.5
Sphinx>=1.3
sphinx-rtd-theme>=0.1
uwsgi>=2.0
Whoosh>=2.6.0
coveralls

